test(d): split test_sources.py -> unit/sources/* (#40)
Seventy tests across 11 classes split by source type into 7 files
under tests/unit/sources/:
- test_database_source.py (TestSourceDataclasses)
- test_file_source.py (TestFileSource + TestFileSourceFromYaml +
  TestFileSourceValidateSourceTable + TestFileSourcesRejectDbFields)
- test_duckdb_file.py (TestDuckDBFileSource)
- test_csv.py (TestCsvSource + merged 1 test from Task D1's test_csv_glob.py)
- test_sqlite.py (TestSqliteSource)
- test_registry.py (TestSourceRegistry + TestLazyRegistry)
- test_protocol.py (TestSourceProtocol)

Classes preserved (Rule M2 exception — they group naturally by source
type and often share fixtures). Source file deleted.
